requirements.txt
flask
groq
python-dotenv
streamlit
PyPDF2